+++ FirebaseInAppMessaging/Sources/Data/FIRIAMMessageContentDataWithImageURL.m
@@ -22,6 +22,7 @@
 #import "FirebaseInAppMessaging/Sources/FIRCore+InAppMessaging.h"
 #import "FirebaseInAppMessaging/Sources/Private/Data/FIRIAMMessageContentData.h"
 #import "FirebaseInAppMessaging/Sources/Private/Data/FIRIAMMessageContentDataWithImageURL.h"
+#import "FirebaseInAppMessaging/Sources/Public/FirebaseInAppMessaging/FIRInAppMessagingErrors.h"
 #import "FirebaseInAppMessaging/Sources/Runtime/FIRIAMSDKRuntimeErrorCodes.h"
 
 static NSInteger const SuccessHTTPStatusCode = 200;
@@ -170,7 +171,7 @@ static NSInteger const SuccessHTTPStatusCode = 200;
                   FIRLogWarning(kFIRLoggerInAppMessaging, @"I-IAM000004", @"%@", errorDesc);
 
                   NSError *error =
-                      [NSError errorWithDomain:kFirebaseInAppMessagingErrorDomain
+                      [NSError errorWithDomain:FIRInAppMessagingErrorDomain
                                           code:FIRIAMSDKRuntimeErrorNonImageMimetypeFromImageURL
                                       userInfo:@{NSLocalizedDescriptionKey : errorDesc}];
                   block(nil, error);