+++ FirebaseInAppMessaging/Tests/Integration/DefaultUITestApp/FiamDisplaySwiftExample/ModalMessageViewController.swift
@@ -48,7 +48,7 @@ class ModalMessageViewController: CommonMessageTestVC {
   }
 
   override func messageDismissed(_ inAppMessage: InAppMessagingDisplayMessage,
-                                 dismissType: FIRInAppMessagingDismissType) {
+                                 dismissType: InAppMessagingDismissType) {
     super.messageDismissed(inAppMessage, dismissType: dismissType)
     verifyLabel.text = "message dismissed!"
   }